requirements.txt
langchain
python-dotenv
langchain-openai
langchain-core
ipython
pydantic
langchain-community
faiss-cpu
requests
langchain-mcp-adapters
fastmcp
